requirements.txt
streamlit
tensorflow-intel==2.12.0
opencv-python==4.6.0.66
numpy==1.23.5
matplotlib
Pillow

